split controls into their own components
