+++ packages/react/src/VideoPlayer/components/controls/CCButton/index.ts
@@ -1,0 +1,1 @@
+export * from './CCButton'
