+++ packages/react/src/VideoPlayer/components/controls/FullScreenButton/index.ts
@@ -1,0 +1,1 @@
+export * from './FullScreenButton'
