+++ packages/react/src/VideoPlayer/components/controls/IconControl/IconControl.tsx
@@ -1,13 +1,13 @@
 import React from 'react'
 import clsx from 'clsx'
-import {Tooltip} from '../index'
+import {Tooltip} from '../../index'
 
-import styles from '../../VideoPlayer.module.css'
+import styles from '../../../VideoPlayer.module.css'
 
 type IconControlProps = {
   tooltip: string
 } & React.HTMLAttributes<HTMLButtonElement>
-
+// TODO Can this be improved?
 export const IconControl = ({tooltip, children, className, ...rest}: IconControlProps) => {
   return (
     <button className={clsx(styles.VideoPlayer__iconControl, className)} {...rest} aria-label={tooltip}>
