+++ packages/react/src/VideoPlayer/components/controls/IconControl/index.ts
