+++ packages/react/src/VideoPlayer/components/controls/MuteButton/index.ts
@@ -1,0 +1,1 @@
+export * from './MuteButton'
