+++ packages/react/src/VideoPlayer/components/controls/PlayPauseButton/index.ts
@@ -1,0 +1,1 @@
+export * from './PlayPauseButton'
