+++ packages/react/src/VideoPlayer/components/controls/VolumeControl/index.ts
@@ -1,0 +1,1 @@
+export * from './VolumeControl'
