+++ packages/react/src/VideoPlayer/components/index.ts
@@ -2,4 +2,3 @@ export * from './captions'
 export * from './controls'
 export * from './range'
 export * from './tooltip'
-export * from './icon-control'
